@@ -4,9 +4,9 @@
     "type": "bundle",
     "require": {
         "php": ">=5.6",
-        "ruflin/elastica": "^3.1|^5.0",
-        "symfony/framework-bundle": "^2.3|^3.0",
-        "symfony/expression-language": "^2.3|^3.0",
+        "ruflin/elastica": "^3.2|^5.0",
+        "symfony/framework-bundle": "^2.7|^3.0",
+        "symfony/expression-language": "^2.7|^3.0",
         "psr/log": "^1.0"
     },
     "require-dev": {
