@@ -3,50 +3,76 @@
 namespace Tests\GBProd\ElasticaBundle\Logger;
 
 use GBProd\ElasticaBundle\Logger\ElasticaLogger;
+use Psr\Log\LoggerInterface;
 
 /**
- * @author Richard Miller <info@limethinking.co.uk>
+ * Tests for ElasticaLogger
+ *
+ * @author GBProd <contact@gb-prod.fr>
  */
 class ElasticaLoggerTest extends \PHPUnit_Framework_TestCase
 {
     /**
-     * @return \PHPUnit_Framework_MockObject_MockObject|\Symfony\Component\HttpKernel\Log\LoggerInterface
+     * @dataProvider getLevels
      */
-    private function getMockLogger()
+    public function testDelegateLogsLevels($level)
     {
-        return $this->getMockBuilder('Psr\Log\LoggerInterface')
-            ->disableOriginalConstructor()
-            ->getMock();
+        $logger = $this->prophesize(LoggerInterface::class);
+
+        $logger
+            ->{$level}('message', ['context'])
+            ->shouldBeCalled()
+        ;
+
+        $testedInstance = new ElasticaLogger($logger->reveal());
+
+        $testedInstance->{$level}('message', ['context']);
+    }
+
+    public function getLevels()
+    {
+        return [
+            ['debug'],
+            ['emergency'],
+            ['alert'],
+            ['critical'],
+            ['error'],
+            ['warning'],
+            ['notice'],
+            ['info'],
+        ];
+    }
+
+    public function testDelegateLog()
+    {
+        $logger = $this->prophesize(LoggerInterface::class);
+
+        $logger
+            ->log('debug', 'message', ['context'])
+            ->shouldBeCalled()
+        ;
+
+        $testedInstance = new ElasticaLogger($logger->reveal());
+
+        $testedInstance->log('debug', 'message', ['context']);
     }
 
     /**
-     * @param string $level
-     * @param string $message
-     * @param array  $context
-     *
-     * @return ElasticaLogger
+     * @dataProvider getLevels
      */
-    private function getMockLoggerForLevelMessageAndContext($level, $message, $context)
+    public function testNotDelegateLogsLevels($level)
     {
-        $loggerMock = $this->getMockBuilder('Psr\Log\LoggerInterface')
-            ->disableOriginalConstructor()
-            ->getMock();
+        $testedInstance = new ElasticaLogger(null);
 
-        $loggerMock->expects($this->once())
-            ->method($level)
-            ->with(
-                $this->equalTo($message),
-                $this->equalTo($context)
-            );
+        // AssertNoErrorExpected
 
-        $elasticaLogger = new ElasticaLogger($loggerMock);
-
-        return $elasticaLogger;
+        $testedInstance->{$level}('message', ['context']);
     }
 
     public function testGetZeroIfNoQueriesAdded()
     {
         $elasticaLogger = new ElasticaLogger();
+
         $this->assertEquals(0, $elasticaLogger->getNbQueries());
     }
 
@@ -54,142 +80,40 @@ class ElasticaLoggerTest extends \PHPUnit_Framework_TestCase
     {
         $elasticaLogger = new ElasticaLogger(null, true);
 
-        $total = rand(1, 15);
-        for ($i = 0; $i < $total; $i++) {
-            $elasticaLogger->logQuery('testPath', 'testMethod', array('data'), 12);
+        for ($i = 0; $i < 15; $i++) {
+            $elasticaLogger->debug('Elastica Request', ['context']);
         }
 
-        $this->assertEquals($total, $elasticaLogger->getNbQueries());
+        $this->assertEquals(15, $elasticaLogger->getNbQueries());
     }
 
-    public function testCorrectlyFormattedQueryReturned()
+    public function testGetQueries()
     {
         $elasticaLogger = new ElasticaLogger(null, true);
 
-        $path   = 'testPath';
-        $method = 'testMethod';
-        $data   = array('data');
-        $time   = 12;
-        $connection = array('host' => 'localhost', 'port' => '8999', 'transport' => 'https');
-        $query = array('search_type' => 'dfs_query_then_fetch');
-
-        $expected = array(
-            'path'        => $path,
-            'method'      => $method,
-            'data'        => $data,
-            'executionMS' => $time,
-            'connection'  => $connection,
-            'queryString' => $query,
-            'engineMS' => 0,
-            'itemCount' => 0
-        );
-
-        $elasticaLogger->logQuery($path, $method, $data, $time, $connection, $query);
+        $elasticaLogger->debug('Elastica Request', ['context']);
         $returnedQueries = $elasticaLogger->getQueries();
-        $this->assertArrayHasKey('backtrace', $returnedQueries[0]);
-        $this->assertNotEmpty($returnedQueries[0]['backtrace']);
-        unset($returnedQueries[0]['backtrace']);
-        $this->assertEquals($expected, $returnedQueries[0]);
+
+        $this->assertEquals(['context'], $returnedQueries[0]);
     }
 
-    public function testNoQueriesStoredIfDebugFalseAdded()
+    public function testNoQueriesStoredIfNoDebug()
     {
         $elasticaLogger = new ElasticaLogger(null, false);
 
-        $total = rand(1, 15);
-        for ($i = 0; $i < $total; $i++) {
-            $elasticaLogger->logQuery('testPath', 'testMethod', array('data'), 12);
+        for ($i = 0; $i < 15; $i++) {
+            $elasticaLogger->debug('Elastica Request', ['context']);
         }
 
         $this->assertEquals(0, $elasticaLogger->getNbQueries());
     }
 
-    public function testQueryIsLogged()
+     public function testNotDelegateLog()
     {
-        $loggerMock = $this->getMockLogger();
+        $testedInstance = new ElasticaLogger(null);
 
-        $elasticaLogger = new ElasticaLogger($loggerMock);
+        // AssertNoErrorExpected
 
-        $path   = 'testPath';
-        $method = 'testMethod';
-        $data   = array('data');
-        $time   = 12;
-
-        $expectedMessage = 'testPath (testMethod) 12000.00 ms';
-
-        $loggerMock->expects($this->once())
-            ->method('info')
-            ->with(
-                $this->equalTo($expectedMessage),
-                $this->equalTo($data)
-            );
-
-        $elasticaLogger->logQuery($path, $method, $data, $time);
-    }
-
-    /**
-     * @return array
-     */
-    public function logLevels()
-    {
-        return array(
-            array('emergency'),
-            array('alert'),
-            array('critical'),
-            array('error'),
-            array('warning'),
-            array('notice'),
-            array('info'),
-            array('debug'),
-        );
-    }
-
-    /**
-     * @dataProvider logLevels
-     */
-    public function testMessagesCanBeLoggedAtSpecificLogLevels($level)
-    {
-        $message = 'foo';
-        $context = array('data');
-
-        $loggerMock = $this->getMockLoggerForLevelMessageAndContext($level, $message, $context);
-
-        call_user_func(array($loggerMock, $level), $message, $context);
-    }
-
-    public function testMessagesCanBeLoggedToArbitraryLevels()
-    {
-        $loggerMock = $this->getMockLogger();
-
-        $level = 'info';
-        $message = 'foo';
-        $context = array('data');
-
-        $loggerMock->expects($this->once())
-            ->method('log')
-            ->with(
-                $level,
-                $message,
-                $context
-            );
-
-        $elasticaLogger = new ElasticaLogger($loggerMock);
-
-        $elasticaLogger->log($level, $message, $context);
-    }
-
-    /**
-     * @dataProvider logLevels
-     */
-    public function testMessagesAreNotLoggedIfNoLogger($level)
-    {
-        $message = 'foo';
-        $context = array('data');
-
-        $elasticaLogger = new ElasticaLogger(null);
-
-        $this->assertNull(
-            call_user_func(array($elasticaLogger, $level), $message, $context)
-        );
+        $testedInstance->log('debug', 'message', ['context']);
     }
 }
